requirements.txt
absl-py==2.2.2 ; python_full_version == "3.12.10"
annotated-types==0.7.0 ; python_full_version == "3.12.10"
asgiref==3.8.1 ; python_full_version == "3.12.10"
astunparse==1.6.3 ; python_full_version == "3.12.10"
attrs==25.3.0 ; python_full_version == "3.12.10"
autobahn==24.4.2 ; python_full_version == "3.12.10"
automat==25.4.16 ; python_full_version == "3.12.10"
certifi==2025.1.31 ; python_full_version == "3.12.10"
cffi==1.17.1 ; python_full_version == "3.12.10" and platform_python_implementation != "PyPy"
channels-redis==4.2.1 ; python_full_version == "3.12.10"
channels==4.2.2 ; python_full_version == "3.12.10"
charset-normalizer==3.4.1 ; python_full_version == "3.12.10"
constantly==23.10.4 ; python_full_version == "3.12.10"
cryptography==44.0.2 ; python_full_version == "3.12.10"
daphne==4.1.2 ; python_full_version == "3.12.10"
django-cors-headers==4.7.0 ; python_full_version == "3.12.10"
django-sslserver==0.22 ; python_full_version == "3.12.10"
django==5.2 ; python_full_version == "3.12.10"
djangorestframework==3.16.0 ; python_full_version == "3.12.10"
dnspython==2.7.0 ; python_full_version == "3.12.10"
filelock==3.18.0 ; python_full_version == "3.12.10"
flatbuffers==25.2.10 ; python_full_version == "3.12.10"
fsspec==2025.3.2 ; python_full_version == "3.12.10"
gast==0.6.0 ; python_full_version == "3.12.10"
google-pasta==0.2.0 ; python_full_version == "3.12.10"
grpcio==1.71.0 ; python_full_version == "3.12.10"
gunicorn==23.0.0 ; python_full_version == "3.12.10"
h5py==3.13.0 ; python_full_version == "3.12.10"
hyperlink==21.0.0 ; python_full_version == "3.12.10"
idna==3.10 ; python_full_version == "3.12.10"
incremental==24.7.2 ; python_full_version == "3.12.10"
jinja2==3.1.6 ; python_full_version == "3.12.10"
keras==3.9.2 ; python_full_version == "3.12.10"
libclang==18.1.1 ; python_full_version == "3.12.10"
markdown-it-py==3.0.0 ; python_full_version == "3.12.10"
markdown==3.8 ; python_full_version == "3.12.10"
markupsafe==3.0.2 ; python_full_version == "3.12.10"
mdurl==0.1.2 ; python_full_version == "3.12.10"
ml-dtypes==0.3.1 ; python_full_version == "3.12.10"
mpmath==1.3.0 ; python_full_version == "3.12.10"
msgpack==1.1.0 ; python_full_version == "3.12.10"
namex==0.0.9 ; python_full_version == "3.12.10"
networkx==3.4.2 ; python_full_version == "3.12.10"
numpy==1.26.0 ; python_full_version == "3.12.10"
nvidia-cublas-cu12==12.4.5.8 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cuda-cupti-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cuda-runtime-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cudnn-cu12==9.1.0.70 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cufft-cu12==11.2.1.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-curand-cu12==10.3.5.147 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cusolver-cu12==11.6.1.9 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cusparse-cu12==12.3.1.170 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-cusparselt-cu12==0.6.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-nccl-cu12==2.21.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-nvjitlink-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
nvidia-nvtx-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
opencv-python==4.11.0.86 ; python_full_version == "3.12.10"
opt-einsum==3.4.0 ; python_full_version == "3.12.10"
optree==0.15.0 ; python_full_version == "3.12.10"
packaging==24.2 ; python_full_version == "3.12.10"
pillow==11.2.1 ; python_full_version == "3.12.10"
protobuf==4.24.1 ; python_full_version == "3.12.10"
psycopg==3.2.6 ; python_full_version == "3.12.10"
pyasn1-modules==0.4.2 ; python_full_version == "3.12.10"
pyasn1==0.6.1 ; python_full_version == "3.12.10"
pycparser==2.22 ; python_full_version == "3.12.10" and platform_python_implementation != "PyPy"
pydantic-core==2.33.1 ; python_full_version == "3.12.10"
pydantic==2.11.3 ; python_full_version == "3.12.10"
pygments==2.19.1 ; python_full_version == "3.12.10"
pymongo==4.12.0 ; python_full_version == "3.12.10"
pyopenssl==25.0.0 ; python_full_version == "3.12.10"
pyyaml==6.0.2 ; python_full_version == "3.12.10"
redis==5.2.1 ; python_full_version == "3.12.10"
requests==2.32.3 ; python_full_version == "3.12.10"
rich==14.0.0 ; python_full_version == "3.12.10"
scipy==1.15.2 ; python_full_version == "3.12.10"
service-identity==24.2.0 ; python_full_version == "3.12.10"
setuptools==79.0.0 ; python_full_version == "3.12.10"
six==1.17.0 ; python_full_version == "3.12.10"
sqlparse==0.5.3 ; python_full_version == "3.12.10"
sympy==1.13.1 ; python_full_version == "3.12.10"
tensorboard-data-server==0.7.2 ; python_full_version == "3.12.10"
tensorboard==2.16.2 ; python_full_version == "3.12.10"
tensorflow==2.16.2 ; python_full_version == "3.12.10"
termcolor==3.0.1 ; python_full_version == "3.12.10"
torch==2.2.2 ; python_full_version == "3.12.10"
torchvision==0.17.2 ; python_full_version == "3.12.10"
triton==3.2.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_full_version == "3.12.10"
twisted==24.11.0 ; python_full_version == "3.12.10"
txaio==23.1.1 ; python_full_version == "3.12.10"
typing-extensions==4.13.2 ; python_full_version == "3.12.10"
typing-inspection==0.4.0 ; python_full_version == "3.12.10"
tzdata==2025.2 ; python_full_version == "3.12.10" and sys_platform == "win32"
urllib3==2.4.0 ; python_full_version == "3.12.10"
werkzeug==3.1.3 ; python_full_version == "3.12.10"
wheel==0.45.1 ; python_full_version == "3.12.10"
wrapt==1.17.2 ; python_full_version == "3.12.10"
zope-interface==7.2 ; python_full_version == "3.12.10"
